www/conoha/requirements.txt
certifi==2019.3.9
chardet==3.0.4
entrypoints==0.3
flake8==3.7.7
future==0.17.1
idna==2.8
jedi==0.13.3
mccabe==0.6.1
parso==0.4.0
pluggy==0.9.0
pycodestyle==2.5.0
pyflakes==2.1.1
python-jsonrpc-server==0.1.2
python-language-server==0.26.1
PyYAML==5.1
requests==2.21.0
texttable==1.6.1
urllib3==1.24.2
